Time: 3 ms (41.54%), Space: 43 MB (84.61%) - LeetHub

+++ 0389-find-the-difference/0389-find-the-difference.java
@@ -3,10 +3,10 @@ class Solution {
         int sum1=0;
         int sum2=0;
         for(int i=0;i<s.length();i++){
-            sum1+=s.charAt(i);
+            sum1+=(int)s.charAt(i);
         }
         for(int i=0;i<t.length();i++){
-            sum2+=t.charAt(i);
+            sum2+=(int)t.charAt(i);
         }
         int res=sum2-sum1;
         return (char) res;
